[ALPS02373903] trusty: fix right shift count >= width of type error.
[Detail]
phys_addr_t >> 32 will cause build error in 32bit kernel because of
phys_addr_t == u32.

[Solution]
Cast phys_addr_t to u64 before right shift.

MTK-Commit-Id: 6b4b290950f8f16b39f9ff80db5b2b2b28a30699

Feature:Google Trusty
CR-Id: ALPS02373903

+++ drivers/trusty/trusty-log.c
@@ -189,7 +189,7 @@ static int trusty_log_probe(struct platform_device *pdev)
 	pa = page_to_phys(s->log_pages);
 	result = trusty_std_call32(s->trusty_dev,
 				   SMC_SC_SHARED_LOG_ADD,
-				   (u32)(pa), (u32)(pa >> 32),
+				   (u32)(pa), (u32)((u64)pa >> 32),
 				   TRUSTY_LOG_SIZE);
 	if (result < 0) {
 		pr_err("trusty std call (SMC_SC_SHARED_LOG_ADD) failed: %d %pa\n",
@@ -222,7 +222,7 @@ error_panic_notifier:
 	trusty_call_notifier_unregister(s->trusty_dev, &s->call_notifier);
 error_call_notifier:
 	trusty_std_call32(s->trusty_dev, SMC_SC_SHARED_LOG_RM,
-			  (u32)pa, (u32)(pa >> 32), 0);
+			  (u32)pa, (u32)((u64)pa >> 32), 0);
 error_std_call:
 	__free_pages(s->log_pages, get_order(TRUSTY_LOG_SIZE));
 error_alloc_log:
@@ -244,7 +244,7 @@ static int trusty_log_remove(struct platform_device *pdev)
 	trusty_call_notifier_unregister(s->trusty_dev, &s->call_notifier);
 
 	result = trusty_std_call32(s->trusty_dev, SMC_SC_SHARED_LOG_RM,
-				   (u32)pa, (u32)(pa >> 32), 0);
+				   (u32)pa, (u32)((u64)pa >> 32), 0);
 	if (result) {
 		pr_err("trusty std call (SMC_SC_SHARED_LOG_RM) failed: %d\n",
 		       result);
